+++ src/scene/Interactive.cpp
@@ -207,7 +207,7 @@ void ARX_INTERACTIVE_DestroyDynamicInfo(Entity * io)
 
 	ARX_INTERACTIVE_ForceIOLeaveZone(io, 0);
 
-	for(long i = 0; i < MAX_EQUIPED; i++) {
+	for(size_t i = 0; i < MAX_EQUIPED; i++) {
 		if(player.equiped[i] == n && ValidIONum(player.equiped[i])) {
 			ARX_EQUIPMENT_UnEquip(entities.player(), entities[player.equiped[i]], 1);
 			player.equiped[i] = EntityHandle();
@@ -498,7 +498,7 @@ void PrepareIOTreatZone(long flag) {
 
 	short sGlobalPlayerRoom = checked_range_cast<short>(PlayerRoom);
 
-	for(long i = 0; i < MAX_EQUIPED; i++) {
+	for(size_t i = 0; i < MAX_EQUIPED; i++) {
 		if(ValidIONum(player.equiped[i])) {
 			Entity *toequip = entities[player.equiped[i]];
 
@@ -1975,7 +1975,7 @@ bool IsEquipedByPlayer(const Entity * io)
 
 	EntityHandle num = io->index();
 
-	for(long i = 0; i < MAX_EQUIPED; i++) {
+	for(size_t i = 0; i < MAX_EQUIPED; i++) {
 		if(ValidIONum(player.equiped[i]) && player.equiped[i] == num)
 			return true;
 	}
